feat: Change button from ArteFinalScreen to BackofficeScreen, and Add another Button with out Tiny integration
